[feature]&[refactor] Install OpenShift spokes with Cilium
The same `openshift-install` command is being invoked directly by Ansible for “hub” clusters, and by some operator for “spoke” clusters. In order to tweak the install-time “shape” of cluster, hub clusters feed off a configuration directory on the bastion VM, while the operator for the spokes consumes a `ConfigMap` Kubernetes object.

- New variable `openshift_install_custom_manifests`, in top-level vars file `vars/openshift-install-custom-manifests.yml` (the intent being to consume it both for hub and spoke clusters)
- New `read_dir_struct` filter
- Piece together the `ConfigMap` object from a mixture of the Git OLM sources, some literal templates, and the occasional dirty search-n'replace, as per the instructions in the [Cilium install documentation](https://docs.cilium.io/en/stable/installation/k8s-install-openshift-okd/)