[test] Prevent `xaasctl` from doing stupid things to kubeconfigs
Before one can log in to a spoke cluster, one should log in to the hub. Unfortunately, it is a good idea to do so from the default kubeconfig (`~/.kube/config`); but a bad idea when `$KUBECONFIG` is set to one of the files in `xaasible-state-cache`.

In this commit, put up a molly-guard in case we do that; this currently happens because of the recursive invocation of `log_in_to_cluster` as the second line of `log_in_to_cluster__spoke`.